add termion, split out modules
