+++ Cargo.toml
@@ -10,7 +10,7 @@ keywords = ["execute", "process", "command", "cli", "interactive"]
 repository = "https://github.com/michaelblawrence/together-rs"
 
 [[bin]]
-name = "together"
+name = "together_tmp"
 path = "src/main.rs"
 
 [dependencies]
@@ -22,4 +22,5 @@ libc = "0.2.153"
 semver = "1.0.22"
 serde = { version = "1.0.196", features = ["derive"] }
 subprocess = "0.2.9"
+termion = "4.0.2"
 toml = "0.8.10"
